[REF] web_tour: remove saveAs attribute
In this commit, in order to simplify the tours API, we remove the saveAs
attribute from the tour object definition. This attribute does not bring
anything in the sense that it just allows not to create in the registry
a tour that could already exist and to save it under another name
to have access to it from tour_service.
To avoid the error of creating a key that already exists in the
registry, simply remove it before adding it again.
This is only used in the website module for the homepage tour.

X-original-commit: e609cfc8237c2630fc018a0d653020a08cc40d5c
